Reorganize and refactor server auth code
