+++ tests/fixtures/dispatcher.py
@@ -2,7 +2,7 @@ import pytest
 from aiohttp import web
 
 from virtool.dispatcher import Connection
-from virtool.app_auth import Client
+from virtool.http.auth import Client
 
 
 class MockWS:
