+++ virtool/app.py
@@ -14,7 +14,7 @@ import pymongo.errors
 from aiohttp import client, web
 from motor import motor_asyncio
 
-import virtool.app_auth
+import virtool.http.auth
 import virtool.app_routes
 import virtool.db.hmm
 import virtool.db.iface
@@ -328,7 +328,7 @@ def create_app(
     ]
 
     if skip_setup:
-        middlewares.append(virtool.app_auth.middleware)
+        middlewares.append(virtool.http.auth.middleware)
 
     app = web.Application(middlewares=middlewares)
